# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Satori R1.1 API
  version: 1.1.0
  description: R1.0 API plus the R1.1 commerce, entitlement, consumption and membership candidate contract frozen by ADR-0003.
servers:
  - url: https://test-satori.shenxinyou.com/api/v1
    description: R1.0 test environment
  - url: http://localhost:3000/api/v1
    description: Local API
...
paths:
...
  /auth/sms-challenges:
    post:
      tags: [Auth]
      operationId: createSmsChallenge
      security: []
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/CreateSmsChallengeRequest'}}}
      responses:
        '202': {description: Challenge accepted, headers: {Retry-After: {schema: {type: integer}}}, content: {application/json: {schema: {$ref: '#/components/schemas/SmsChallengeEnvelope'}}}}
        default: {$ref: '#/components/responses/Error'}
...
  /me:
    get:
      tags: [Me]
      operationId: getMe
      responses:
        '200': {description: Current user routing state, content: {application/json: {schema: {$ref: '#/components/schemas/MeEnvelope'}}}}
        default: {$ref: '#/components/responses/Error'}
...
  /me/home-overview:
    get:
      tags: [Me]
      operationId: getHomeOverview
      responses:
        '200': {description: Stable R1.0 home projection, content: {application/json: {schema: {$ref: '#/components/schemas/HomeOverviewEnvelope'}}}}
        default: {$ref: '#/components/responses/Error'}
...
  /daily-insights/today:
    post:
      tags: [DailyInsight]
      operationId: createOrGetTodayDailyInsight
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false}}}}
      responses:
        '200': {description: Existing ready insight, content: {application/json: {schema: {$ref: '#/components/schemas/DailyInsightCommandEnvelope'}}}}
        '202': {description: Generation accepted, content: {application/json: {schema: {$ref: '#/components/schemas/DailyInsightCommandEnvelope'}}}}
        default: {$ref: '#/components/responses/Error'}
  /daily-insights:
    get:
      tags: [DailyInsight]
      operationId: listDailyInsights
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses:
        '200': {description: Daily insight history, content: {application/json: {schema: {$ref: '#/components/schemas/DailyInsightListEnvelope'}}}}
        default: {$ref: '#/components/responses/Error'}
  /daily-insights/{localDate}:
    get:
      tags: [DailyInsight]
      operationId: getDailyInsightByLocalDate
      parameters:
        - {name: localDate, in: path, required: true, schema: {type: string, format: date}}
      responses:
        '200': {description: Insight for local date, content: {application/json: {schema: {$ref: '#/components/schemas/DailyInsightEnvelope'}}}}
        default: {$ref: '#/components/responses/Error'}
...
  /me/entitlements:
    get:
      tags: [Entitlements]
      operationId: listEntitlements
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses:
        '200': {description: Source-separated entitlement grants, content: {application/json: {schema: {$ref: '#/components/schemas/EntitlementListEnvelope'}}}}
        default: {$ref: '#/components/responses/Error'}
  /me/entitlements/{entitlementId}:
    get:
      tags: [Entitlements]
      operationId: getEntitlement
      parameters: [{$ref: '#/components/parameters/EntitlementId'}]
      responses:
        '200': {description: Entitlement grant detail, content: {application/json: {schema: {$ref: '#/components/schemas/EntitlementEnvelope'}}}}
        default: {$ref: '#/components/responses/Error'}
  /me/usage-records:
    get:
      tags: [Entitlements]
      operationId: listUsageRecords
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses:
        '200': {description: Entitlement usage records, content: {application/json: {schema: {$ref: '#/components/schemas/UsageRecordListEnvelope'}}}}
        default: {$ref: '#/components/responses/Error'}
  /entitlement-resolutions:
    post:
      tags: [Consumption]
      operationId: createEntitlementResolution
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateEntitlementResolutionRequest'}}}}
      responses:
        '201': {description: System-only fixed-priority resolution, content: {application/json: {schema: {$ref: '#/components/schemas/EntitlementResolutionEnvelope'}}}}
        default: {$ref: '#/components/responses/Error'}
  /consumption-intents:
    post:
      tags: [Consumption]
      operationId: createConsumptionIntent
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateConsumptionIntentRequest'}}}}
      responses:
        '201': {description: Selected source reserved for 30 minutes before draw, content: {application/json: {schema: {$ref: '#/components/schemas/ConsumptionIntentEnvelope'}}}}
        default: {$ref: '#/components/responses/Error'}
  /consumption-intents/{intentId}:
    get:
      tags: [Consumption]
      operationId: getConsumptionIntent
      parameters: [{$ref: '#/components/parameters/ConsumptionIntentId'}]
      responses:
        '200': {description: Consumption intent status, content: {application/json: {schema: {$ref: '#/components/schemas/ConsumptionIntentEnvelope'}}}}
        default: {$ref: '#/components/responses/Error'}
  /consumption-intents/{intentId}/start:
    post:
      tags: [Consumption]
      operationId: startConsumptionIntent
      parameters: [{$ref: '#/components/parameters/ConsumptionIntentId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses:
        '200': {description: Formal business execution started, content: {application/json: {schema: {$ref: '#/components/schemas/ConsumptionIntentEnvelope'}}}}
        default: {$ref: '#/components/responses/Error'}
...
  /feedback:
    post:
      tags: [Me]
      operationId: createFeedback
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateFeedbackRequest'}}}}
      responses:
        '201': {description: Feedback created, content: {application/json: {schema: {$ref: '#/components/schemas/FeedbackEnvelope'}}}}
        default: {$ref: '#/components/responses/Error'}
  /me/account-deletion-requests:
    post:
      tags: [Me]
      operationId: createAccountDeletionRequest
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateAccountDeletionRequest'}}}}
      responses:
        '202': {description: Deletion accepted, content: {application/json: {schema: {$ref: '#/components/schemas/AccountDeletionEnvelope'}}}}
        default: {$ref: '#/components/responses/Error'}
  /me/account-deletion-request:
    get:
      tags: [Me]
      operationId: getAccountDeletionRequest
      responses:
        '200': {description: Current deletion request, content: {application/json: {schema: {$ref: '#/components/schemas/AccountDeletionEnvelope'}}}}
        default: {$ref: '#/components/responses/Error'}
    delete:
      tags: [Me]
      operationId: cancelAccountDeletionRequest
      responses:
        '204': {description: Deletion request cancelled}
        default: {$ref: '#/components/responses/Error'}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
    refreshCookie: {type: apiKey, in: cookie, name: satori_refresh}
  parameters:
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
    Cursor: {name: cursor, in: query, required: false, schema: {type: string}}
    Limit: {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 50, default: 20}}
...
    EntitlementId: {name: entitlementId, in: path, required: true, schema: {type: string}}
    ConsumptionIntentId: {name: intentId, in: path, required: true, schema: {type: string}}
...
  responses:
    Error:
      description: API v1 error
      headers: {X-Request-Id: {schema: {type: string}}}
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties: {code: {type: string}, message: {type: string}, requestId: {type: string}, retryable: {type: boolean}, nextAction: {type: [string, 'null']}, retryAfterSeconds: {type: [integer, 'null'], minimum: 0}, details: {type: [object, 'null'], additionalProperties: true}}
    PageMeta: {type: object, required: [nextCursor, hasMore], properties: {nextCursor: {type: [string, 'null']}, hasMore: {type: boolean}}}
    Preferences: {type: object, required: [timezone, locale], properties: {timezone: {type: string, examples: [Asia/Shanghai]}, timezoneSource: {type: string}, locale: {type: string, examples: [zh-CN]}, updatedAt: {type: string, format: date-time}}}
...
    LegalDocumentSummary: {type: object, required: [documentId, type, version, title, required], properties: {documentId: {type: string}, type: {type: string, enum: [PRIVACY_POLICY, TERMS_OF_SERVICE, AI_CONTENT_NOTICE]}, version: {type: string}, title: {type: string}, required: {type: boolean}, publishedAt: {type: string, format: date-time}}}
...
    CreateSmsChallengeRequest:
      type: object
      required: [phone, purpose, device]
      properties:
        phone: {type: object, required: [countryCode, nationalNumber], properties: {countryCode: {type: string}, nationalNumber: {type: string}}}
        purpose: {type: string, enum: [LOGIN, ACCOUNT_DELETION, SECURITY_CONFIRMATION]}
        device: {$ref: '#/components/schemas/Device'}
    Device: {type: object, required: [deviceId, timezone], properties: {deviceId: {type: string}, timezone: {type: string}, appVersion: {type: string}}}
    SmsChallengeEnvelope: {type: object, required: [data], properties: {data: {type: object, required: [challengeId, expiresAt, resendAvailableAt, phoneMasked], properties: {challengeId: {type: string}, expiresAt: {type: string, format: date-time}, resendAvailableAt: {type: string, format: date-time}, phoneMasked: {type: string}}}}}
...
    UserSummary: {type: object, required: [userId, status, phoneMasked, requiresConsent, createdAt], properties: {userId: {type: string}, status: {type: string, enum: [ACTIVE, DELETION_PENDING, DELETING, DISABLED]}, phoneMasked: {type: string}, requiresConsent: {type: boolean}, createdAt: {type: string, format: date-time}}}
    NextAction: {type: string, enum: [ACCEPT_CONSENTS, CREATE_PROFILE, CONFIRM_PROFILE, CLAIM_REGISTRATION_REWARD, CREATE_TODAY_DAILY_INSIGHT, VIEW_HOME]}
...
    MeEnvelope: {type: object, required: [data], properties: {data: {allOf: [{$ref: '#/components/schemas/UserSummary'}, {type: object, required: [preferences, nextAction], properties: {preferences: {$ref: '#/components/schemas/Preferences'}, pendingConsents: {type: array, items: {$ref: '#/components/schemas/LegalDocumentSummary'}}, profileState: {type: string, enum: [NOT_CREATED, CALCULATED, ACTIVE]}, nextAction: {$ref: '#/components/schemas/NextAction'}}}]}}}
...
    DailyEnergySummary: {type: object, required: [greeting, guidance, energyLevel, suitableActions, cautions, date, dayCard, heavenCard, score, signals, ruleVersion, copyVersion], properties: {greeting: {type: string}, guidance: {type: string}, energyLevel: {type: string, enum: [高, 中, 低]}, suitableActions: {type: array, items: {type: string}}, cautions: {type: array, items: {type: string}}, date: {type: string}, dayCard: {type: string}, heavenCard: {type: string}, score: {type: number}, signals: {type: array, items: {type: string}}, ruleVersion: {type: string}, copyVersion: {type: string}}}
    HomeOverviewEnvelope: {type: object, required: [data], properties: {data: {type: object, required: [user, profile, registrationReward, wisdomSeedAccount, dailyInsight, dailyEnergySummary, nextAction], properties: {user: {type: object, properties: {userId: {type: string}, timezone: {type: string}}}, profile: {$ref: '#/components/schemas/LifeProfile'}, cards: {type: array, items: {$ref: '#/components/schemas/RelationshipCard'}}, registrationReward: {$ref: '#/components/schemas/RegistrationReward'}, wisdomSeedAccount: {$ref: '#/components/schemas/WisdomSeedAccount'}, dailyInsight: {type: object, properties: {localDate: {type: string, format: date}, state: {type: string, enum: [NOT_CREATED, GENERATING, READY, FAILED]}, dailyInsightId: {type: [string, 'null']}, taskId: {type: [string, 'null']}}}, dailyEnergySummary: {type: object, required: [state, data], properties: {state: {type: string, enum: [READY, UNAVAILABLE, NOT_AVAILABLE]}, data: {oneOf: [{$ref: '#/components/schemas/DailyEnergySummary'}, {type: 'null'}]}}}, nextAction: {$ref: '#/components/schemas/NextAction'}}}}}
...
    RelationshipCard:
      type: object
      required: [dimension, title, order, cardId, cardCode, ganzhi, zodiac, season, talentMark, abilityMark, journeyMark, deckCode, deckVersion, assetUrl, altText, summary]
      properties:
        dimension: {type: string, enum: [SPACETIME, CAREER, FAMILY, SELF]}
        title: {type: string}
        order: {type: integer, minimum: 1, maximum: 4}
        cardId: {type: [integer, 'null'], minimum: 1, maximum: 60}
        cardCode: {type: string}
        ganzhi: {type: [string, 'null']}
        zodiac: {type: [string, 'null']}
        season: {type: [string, 'null']}
        talentMark: {type: [string, 'null']}
        abilityMark: {type: [string, 'null']}
        journeyMark: {type: [string, 'null']}
        deckCode: {type: string}
        deckVersion: {type: string}
        assetUrl: {type: [string, 'null']}
        altText: {type: string}
        summary: {type: string}
        uncertainty: {type: [string, 'null']}
        mappingVersion: {type: string}
        knowledgeVersion: {type: string}
    LifeProfile: {type: object, required: [profileId, subjectType, displayName, state], properties: {profileId: {type: string}, subjectId: {type: string}, subjectType: {type: string, enum: [SELF, OTHER]}, displayName: {type: string}, relationshipType: {type: string, enum: [SELF, FAMILY, FRIEND, COLLEAGUE, OTHER]}, groupId: {type: [string, 'null']}, currentRevisionId: {type: [string, 'null']}, pendingRevisionId: {type: [string, 'null']}, state: {type: string, enum: [NOT_CREATED, CALCULATED, ACTIVE, DELETED]}, updatedAt: {type: string, format: date-time}}}
...
    WisdomSeedAccount: {type: object, required: [accountId, available, reserved, totalEarned, totalSpent, updatedAt], properties: {accountId: {type: string}, available: {type: integer, minimum: 0}, reserved: {type: integer, minimum: 0}, totalEarned: {type: integer, minimum: 0}, totalSpent: {type: integer, minimum: 0}, updatedAt: {type: string, format: date-time}}}
...
    RegistrationReward: {type: object, required: [rewardId, rewardType, status, wisdomSeedAmount], properties: {rewardId: {type: string}, rewardType: {const: NEW_USER_ONBOARDING}, status: {type: string, enum: [AVAILABLE, CLAIMED, EXPIRED, INELIGIBLE]}, wisdomSeedAmount: {type: integer, minimum: 0}, claimedAt: {type: [string, 'null'], format: date-time}, expiresAt: {type: [string, 'null'], format: date-time}}}
...
    CommerceServiceType: {type: string, enum: [DAILY_ENERGY, CARD_READING]}
...
    BusinessContext: {type: object, additionalProperties: false, required: [type, id], properties: {type: {type: string, enum: [DAILY_INSIGHT, CARD_READING_INTENT, MEMBERSHIP_RENEWAL, MEMBERSHIP_UPGRADE]}, id: {type: string}}}
...
    EntitlementEnvelope: {type: object, required: [data], properties: {data: {$ref: '#/components/schemas/EntitlementGrant'}}}
    EntitlementListEnvelope: {type: object, required: [data, meta], properties: {data: {type: array, items: {$ref: '#/components/schemas/EntitlementGrant'}}, meta: {$ref: '#/components/schemas/PageMeta'}}}
    UsageRecord: {type: object, additionalProperties: false, required: [recordId, entitlementId, type, quantity, businessContext, createdAt], properties: {recordId: {type: string}, entitlementId: {type: string}, type: {type: string, enum: [GRANT, RESERVE, COMMIT, RELEASE, REVERSE, EXPIRE, FREEZE, UNFREEZE, FORFEIT, ADJUSTMENT]}, quantity: {type: integer}, businessContext: {$ref: '#/components/schemas/BusinessContext'}, originalRecordId: {type: [string, 'null']}, createdAt: {type: string, format: date-time}}}
    UsageRecordListEnvelope: {type: object, required: [data, meta], properties: {data: {type: array, items: {$ref: '#/components/schemas/UsageRecord'}}, meta: {$ref: '#/components/schemas/PageMeta'}}}
    CreateEntitlementResolutionRequest: {type: object, additionalProperties: false, required: [serviceType, quantity, businessContext], properties: {serviceType: {$ref: '#/components/schemas/CommerceServiceType'}, quantity: {type: integer, minimum: 1}, cardCount: {type: [integer, 'null'], minimum: 1, maximum: 5}, businessContext: {$ref: '#/components/schemas/BusinessContext'}}}
...
    EntitlementResolutionEnvelope: {type: object, required: [data], properties: {data: {$ref: '#/components/schemas/EntitlementResolution'}}}
    CreateConsumptionIntentRequest: {type: object, additionalProperties: false, required: [resolutionId], properties: {resolutionId: {type: string}}}
...
    ConsumptionIntentEnvelope: {type: object, required: [data], properties: {data: {$ref: '#/components/schemas/ConsumptionIntent'}}}
...
    GenerationTask: {type: object, required: [taskId, type, status, stage, stageLabel, canRetry, target, createdAt, updatedAt], properties: {taskId: {type: string}, type: {const: DAILY_INSIGHT_GENERATION}, status: {type: string, enum: [PENDING, GENERATING, READY, FAILED]}, stage: {type: string, enum: [QUEUED, PREPARING_PROFILE, PREPARING_CONTEXT, GENERATING_CONTENT, VALIDATING_CONTENT, FINALIZING, COMPLETED, RETRY_WAITING, FAILED]}, stageLabel: {type: string}, canRetry: {type: boolean}, target: {type: object, required: [type, id], properties: {type: {const: DAILY_INSIGHT}, id: {type: string}}}, failure: {type: [object, 'null'], properties: {code: {type: string}, message: {type: string}, retryable: {type: boolean}}}, createdAt: {type: string, format: date-time}, updatedAt: {type: string, format: date-time}, completedAt: {type: [string, 'null'], format: date-time}}}
...
    DailyInsight: {type: object, required: [dailyInsightId, localDate, timezone, status, settlement, createdAt], properties: {dailyInsightId: {type: string}, localDate: {type: string, format: date}, timezone: {type: string}, status: {type: string, enum: [GENERATING, READY, FAILED]}, profileRevisionId: {type: [string, 'null']}, content: {type: [object, 'null'], properties: {theme: {type: string}, insight: {type: string}, action: {type: string}, reflectionQuestion: {type: string}, notice: {type: string}}}, fallback: {type: [object, 'null'], properties: {title: {type: string}, message: {type: string}}}, taskId: {type: [string, 'null']}, settlement: {$ref: '#/components/schemas/SeedSettlement'}, publishedAt: {type: [string, 'null'], format: date-time}, createdAt: {type: string, format: date-time}}}
    DailyInsightEnvelope: {type: object, required: [data], properties: {data: {$ref: '#/components/schemas/DailyInsight'}}}
    DailyInsightCommandEnvelope: {type: object, required: [data], properties: {data: {type: object, required: [dailyInsight, task], properties: {dailyInsight: {$ref: '#/components/schemas/DailyInsight'}, task: {oneOf: [{$ref: '#/components/schemas/GenerationTask'}, {type: 'null'}]}}}}}
    DailyInsightListEnvelope: {type: object, required: [data, meta], properties: {data: {type: array, items: {$ref: '#/components/schemas/DailyInsight'}}, meta: {$ref: '#/components/schemas/PageMeta'}}}
    CreateFeedbackRequest: {type: object, required: [target, rating, reasons], properties: {target: {type: object, required: [type, id], properties: {type: {type: string, enum: [LIFE_REPORT, DAILY_INSIGHT]}, id: {type: string}, sectionCode: {type: [string, 'null']}}}, rating: {type: string, enum: [HELPFUL, NOT_RESONANT, UNCOMFORTABLE]}, reasons: {type: array, items: {type: string, enum: [TOO_ABSOLUTE, NOT_ACCURATE, TOO_GENERIC, HARMFUL_OR_OFFENSIVE, OTHER]}}, comment: {type: [string, 'null'], maxLength: 1000}}}
    FeedbackEnvelope: {type: object, required: [data], properties: {data: {type: object, required: [feedbackId, target, rating, createdAt], properties: {feedbackId: {type: string}, target: {type: object, additionalProperties: true}, rating: {type: string}, createdAt: {type: string, format: date-time}}}}}
    CreateAccountDeletionRequest: {type: object, required: [smsChallengeId, verificationCode, reason], properties: {smsChallengeId: {type: string}, verificationCode: {type: string}, reason: {type: string, enum: [NO_LONGER_NEEDED, PRIVACY_CONCERN, OTHER]}}}
...
    AccountDeletionEnvelope: {type: object, required: [data], properties: {data: {$ref: '#/components/schemas/AccountDeletion'}}}
...
security:
  - bearerAuth: []